corrected folder structure of chat screen page
