+++ frontend/src/App.js
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import LandingPage from './components/LandingPage';
-import ChatScreenPage from './components/ChatScreenPage';
+import ChatScreenPage from './components/ChatScreenPage/ChatScreenPage/ChatScreenPage';
 
 function App() {
   return (
