+++ frontend/src/components/ChatScreenPage/ChatScreenPage/ChatScreenPage.css
@@ -1,0 +1,124 @@
+/* Main container for the chat screen */
+.chat-screen-container {
+    display: flex;
+    height: 100vh;
+    padding: 0 5px;
+    position: relative;
+    background-color: #eaeaea;
+  }
+  
+  /* Hamburger menu for the sidebar or navigation */
+  .hamburger {
+    font-size: 30px;
+    cursor: pointer;
+    position: absolute;
+    top: 20px;
+    right: 30px;
+    z-index: 10;
+    background-color: transparent;
+    border: none;
+  }
+  
+  /* Section to display user's query and source */
+  .question-source-section {
+    display: flex;
+    align-items: center;
+    margin-bottom: 20px;
+  }
+  
+  /* Separator between sections */
+  .question-separator {
+    width: 100%;
+    border-top: 1px solid #ddd;
+    margin: 10px 0;
+  }
+  
+  /* Question display: Blue circle, "You" label, and the user's question */
+  .question-display {
+    display: flex;
+    align-items: center;
+    margin-right: 20px;
+  }
+  
+  /* User's question text display */
+  .question-text {
+    position: relative;
+    background-color: #eae9e9;
+    border-radius: 20px;
+    padding: 10px 20px;
+    color: black;
+    font-size: 18px;
+    max-width: 60%;
+    word-wrap: break-word;
+    margin-left: 10px;
+    border: 1px solid #ddd;
+  }
+  
+  /* Triangle (bubble tail) for the question bubble */
+  .question-text::before {
+    content: ""; /* No text content, just the shape */
+    position: absolute; /* Position relative to the bubble */
+    bottom: 24.5px; /* Aligns with the bottom of the bubble */
+    left: -1px; /* Positioned to the left side */
+    border-width: 15px 15px 0 0; /* Creates a triangle pointing down */
+    border-style: solid; /* Solid lines */
+    border-color: #e3e3e3 transparent transparent transparent; /* Matches bubble background */
+  }
+  
+  /* Section for the results */
+  .results-section {
+    flex: 1;
+    overflow-y: auto;
+  }
+  
+  /* Individual result card styling */
+  .result-card {
+    margin-bottom: 15px;
+    padding: 10px;
+    border: 1px solid #e3e3e3;
+    border-radius: 9px;
+    background-color: #e3e3e3;
+    margin-left: 60px;
+    margin-top: -1px;
+    position: relative;
+  }
+  
+  /* Triangle (bubble tail) for the result card */
+  .result-card::before {
+    content: ""; /* No text content, just the shape */
+    position: absolute; /* Position relative to the card */
+    
+    margin-top: -11px;
+    left: -1px; /* Positioned to the left side */
+    border-width: 15px 15px 0 0; /* Creates a triangle pointing down */
+    border-style: solid; /* Solid lines */
+    border-color:  #e3e3e3 transparent transparent transparent; /* Matches card background */
+  }
+  
+  /* Result bubble styling */
+  .result-bubble {
+    position: relative;
+    background-color: #f3f3f3;
+    border-radius: 20px;
+    padding: 10px 20px;
+    color: black;
+    font-size: 16px;
+    max-width: 60%;
+    word-wrap: break-word;
+    margin-top: 20px;
+    border: 1px solid #ddd;
+  }
+  .zoom-btn {
+    background-color: #4CAF50; /* Green */
+    border: none;
+    color: white;
+    padding: 5px 10px;
+    text-align: center;
+    text-decoration: none;
+    display: inline-block;
+    font-size: 14px;
+    margin-top: 10px;
+    cursor: pointer;
+    border-radius: 4px;
+  }
+
